pyproject.toml
[project]
name = "project-tui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai==1.12.1",
    "textual>=6.3.0",
]
